// src/types/apps/contact.ts
export interface ContactType {
  id: number | string;
  firstname: string;
  lastname: string;
  image: string;
  department: string;
  company: string;
  phone: string;
  email: string;
  password: string;
  token: string;
  address: string;
  notes: string;
  frequentlycontacted: boolean;
  role: string;
  starred: boolean;
  deleted: boolean;
}